Rename action and reducer files.
