+++ dashboard/src/reducers/index.ts
@@ -5,9 +5,9 @@ import { connectRouter } from "connected-react-router";
 import { History } from "history";
 import { combineReducers } from "redux";
 import { IStoreState } from "shared/types";
-import appsReducer from "./apps";
+import appsReducer from "./installedpackages";
 import authReducer from "./auth";
-import packageReducer from "./packages";
+import packageReducer from "./availablepackages";
 import clusterReducer from "./cluster";
 import configReducer from "./config";
 import kubeReducer from "./kube";
